LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-data-url-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: undefined_variable

FAIL window.onerror - runtime error in <script src=data:...> assert_equals: second arg expected "data:text/javascript,undefined_variable;" but got ""
PASS window.onerror - runtime error in <script src=data:...> (column) 
